requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
python-whois
python-dateutil
cryptography